+++ src/shared/world/entity/ai/behavior/BehaviorControl.cs
@@ -4,8 +4,8 @@ public interface BehaviorControl<T> where T : IEntity
 {
     public string name { get; }
     public Status getStatus();
-    public bool tryStart(T agent, float currentTime, float deltaTime);
-    public void updateOrStop(T agent, float currentTime, float deltaTime);
+    public bool tryStart(T agent, double currentTime, double deltaTime);
+    public void updateOrStop(T agent, double currentTime, double deltaTime);
     public void stop(T agent);
 }
